update controller method
